everthing in the custom script arta behaviour is working, but not update(dt)

+++ NDR2DGameEngine/InputManager.cpp
@@ -4,14 +4,15 @@ bool ArtaEngine::InputManager::isSpriteClicked(sf::Sprite object, sf::Mouse::But
 {
     if (sf::Mouse::isButtonPressed(button))
     {
-        sf::IntRect tempRect(object.getPosition().x, object.getPosition().y,
-            object.getGlobalBounds().width, object.getGlobalBounds().height);
+        sf::IntRect tempRect(static_cast<int>(object.getPosition().x), static_cast<int>(object.getPosition().y),
+            static_cast<int>(object.getGlobalBounds().width), static_cast<int>(object.getGlobalBounds().height));
 
         return tempRect.contains(sf::Mouse::getPosition(window));
     }
 
     return false;
 }
+
 
 sf::Vector2i ArtaEngine::InputManager::GetMousePosition(sf::RenderWindow& window)
 {
